Time: 188 ms (75.56%), Space: 79.8 MB (50.62%) - LeetHub

+++ 2616-maximal-score-after-applying-k-operations/2616-maximal-score-after-applying-k-operations.cpp
@@ -14,7 +14,7 @@ public:
 
             ans+=temp;
 
-            int toPush = ceil(double((double)temp/3));
+            int toPush = ceil(((double)temp/3));
             // cout << double((double)temp/3) << endl;
             pq.push(toPush);
